+++ tests/sdk/Scanner/WfpProvider/WfpCalculator/samples/header_py_file.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2024, SCANOSS
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction.
+
+import os
+import sys
+from pathlib import Path
+
+def main():
+    print("Hello, World!")
+
+if __name__ == "__main__":
+    main()